### Summary of Most Important Changes
The codebase has undergone significant refactoring to improve readability, maintainability, and functionality. Key changes include the introduction of new methods for processing sales and calculating taxes, updates to data types for better precision, and the addition of new classes and services for handling currency data and tax bracket information. The `KazancHesapla` class has been simplified, and new methods have been introduced to streamline the calculation logic. The `DataInit` class has been updated to include new stock data and call the new methods for income and tax calculations. The `Alis` and `Satis` classes now use `decimal` for the `Adet` property and fetch currency data using services. The `KarZarar` class has been updated for detailed financial tracking, and the `Sonuc` class has been removed. New classes and services have been added to handle currency data and tax bracket information.

### List of Changes

1. **KazancHesapla.cs**:
   - Introduced a constant `ufeLimit` for UFE calculations.
   - Refactored `UfeHesapla`, `MaliyetHesapla`, and `HesaplaKarZarar` methods to be private and simplified their logic.
   - Added a new method `ProcessSatis` to handle sales processing logic.
   - Refactored `GelirHesapla` method to use `ProcessSatis` for processing sales.
   - Added a new method `VergiHesapla` to calculate taxes based on income.
   - **Reference**: [KazancHesapla.cs]

2. **DataInit.cs**:
   - Added a new stock symbol "3x Long QQQ" with purchase and sale transactions.
   - Changed the type of `sonuclar` from `List<Sonuc>` to `List<KarZarar>`.
   - Updated the method call from `CalculateTax` to `GelirHesapla`.
   - Added a call to `VergiHesapla` to calculate taxes based on the results.
   - **Reference**: [DataInit.cs]

3. **Alis.cs**:
   - Changed the type of `Adet` from `int` to `decimal`.
   - **Reference**: [Alis.cs]

4. **KarZarar.cs**:
   - Removed several properties related to profit/loss calculations.
   - Added new properties for detailed income and expense tracking.
   - Added computed properties for various financial metrics.
   - **Reference**: [KarZarar.cs]

5. **Satis.cs**:
   - Changed the type of `Adet` from `int` to `decimal`.
   - Updated the `Kur` and `Ufe` properties to use `DovizService` and `YufeService`.
   - **Reference**: [Satis.cs]

6. **Sonuc.cs**:
   - Removed the `Sonuc` class entirely.
   - **Reference**: [Sonuc.cs]

7. **Program.cs**:
   - Added references to `YurtDisiKazancHesapla.Business`.
   - Added a call to `VergiHesapla` to calculate taxes based on the results.
   - **Reference**: [Program.cs]

8. **DovizService.cs**:
   - Added references to `System.Text.Json` and `Newtonsoft.Json`.
   - Added a new namespace `YurtDisiKazancHesapla.Models.Service`.
   - **Reference**: [DovizService.cs]

9. **YurtDisiKazancHesapla.csproj**:
   - Added a package reference for `Newtonsoft.Json`.
   - **Reference**: [YurtDisiKazancHesapla.csproj]

10. **HıstoryModel.cs**:
    - Added a new class `HıstoryModel` with nested classes for handling currency data.
    - **Reference**: [HıstoryModel.cs]

11. **Kademeler.cs**:
    - Added a new class `Kademeler` for tax bracket information.
    - **Reference**: [Kademeler.cs]

12. **VergiDilim.cs**:
    - Added a new class `VergiDilim` for tax bracket information.
    - **Reference**: [VergiDilim.cs]

13. **DovizComService.cs**:
    - Added a new service class `DovizComService` for fetching currency data from an external API.
    - Defined enums for `Emtia` and `Aralik` to specify the type of commodity and time range for data fetching.
    - Implemented methods to fetch historical data for USD, ONS, and ALTINS1.
    - **Reference**: [DovizComService.cs]

+++ YurtDisiKazancHesapla/Models/Alis.cs
@@ -16,7 +16,7 @@ public class Alis
 {
     public DateTime Tarih { get; set; }
     public decimal Fiyat { get; set; }
-    public int Adet { get; set; }
+    public decimal Adet { get; set; }
     public decimal ToplamTutar => Fiyat * Adet;
     public decimal Kur => DovizService.GetDovizKuru(Tarih).Result;
     public decimal Ufe => YufeService.GetValueForYearAndMonth(Tarih.Year, Tarih.Month);
